+++ src/languages/esc-pos.js
@@ -125,12 +125,10 @@ class LanguageEscPos {
     /**
      * Generate a QR code
      * @param {string} value        Value to encode
-     * @param {number} model        QR Code model (1 or 2)
-     * @param {number} size         QR Code size (1 to 8)
-     * @param {string} errorlevel   Error correction level ('l', 'm', 'q', 'h')
+     * @param {object} options      Configuration object
      * @returns {Array}             Array of bytes to send to the printer
      */
-    qrcode(value, model, size, errorlevel) {
+    qrcode(value, options) {
         let result = [];
 
         /* Model */
@@ -140,13 +138,9 @@ class LanguageEscPos {
             2: 0x32,
         };
   
-        if (typeof model === 'undefined') {
-            model = 2;
-        }
-  
-        if (model in models) {
+        if (options.model in models) {
             result.push(
-                0x1d, 0x28, 0x6b, 0x04, 0x00, 0x31, 0x41, models[model], 0x00
+                0x1d, 0x28, 0x6b, 0x04, 0x00, 0x31, 0x41, models[options.model], 0x00
             );
         } else {
             throw new Error('Model must be 1 or 2');
@@ -154,20 +148,16 @@ class LanguageEscPos {
   
         /* Size */
   
-        if (typeof size === 'undefined') {
-            size = 6;
-        }
-  
-        if (typeof size !== 'number') {
+        if (typeof options.size !== 'number') {
             throw new Error('Size must be a number');
         }
   
-        if (size < 1 || size > 8) {
+        if (options.size < 1 || options.size > 8) {
             throw new Error('Size must be between 1 and 8');
         }
   
         result.push(
-            0x1d, 0x28, 0x6b, 0x03, 0x00, 0x31, 0x43, size
+            0x1d, 0x28, 0x6b, 0x03, 0x00, 0x31, 0x43, options.size
         );
   
         /* Error level */
@@ -179,16 +169,12 @@ class LanguageEscPos {
             'h': 0x33,
         };
   
-        if (typeof errorlevel === 'undefined') {
-            errorlevel = 'm';
-        }
-  
-        if (errorlevel in errorlevels) {
+        if (options.errorlevel in errorlevels) {
             result.push(
-                0x1d, 0x28, 0x6b, 0x03, 0x00, 0x31, 0x45, errorlevels[errorlevel]
+                0x1d, 0x28, 0x6b, 0x03, 0x00, 0x31, 0x45, errorlevels[options.errorlevel]
             );
         } else {
-                throw new Error('Error level must be l, m, q or h');
+            throw new Error('Error level must be l, m, q or h');
         }
   
         /* Data */